feat: index, App, router, common view typescript 변환
