--- src/index.js
@@ -1,5 +1,0 @@
-import App from './App';
-import 'styles/reset.scss';
-import 'styles/common.scss';
-
-new App(document.querySelector('#root'));
