+++ src/index.ts
@@ -1,0 +1,5 @@
+import App from './App';
+import './styles/reset.scss';
+import './styles/common.scss';
+
+new App(document.querySelector('#root')! as HTMLElement);
